backend/requirements.txt
Flask==3.1.0
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
stripe==12.1.0
cryptography==43.0.1
pytest==8.3.5
flake8==7.2.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
alembic==1.12.1